Fixed 3d evaluations of polynomials and a few other bugs.

+++ fem/fe_coll.cpp
@@ -273,6 +273,7 @@ FiniteElementCollection *FiniteElementCollection::New(const char *name)
    }
    else if (!strncmp(name, "RBF", 3) || !strncmp(name, "RK", 2))
    {
+      // Example: RK4_IM_M_2d_020_4.01
       int dim = atoi(name + 9);
       int numPoints = atoi(name + 12);
       double h = atof(name + 16);
@@ -321,10 +322,6 @@ FiniteElementCollection *FiniteElementCollection::New(const char *name)
                                       func, dist);
       }
    }
-   // else if (!strncmp(name, "RK", 2))
-   // {
-   //    fec = new RK_FECollection();
-   // }
    else
    {
       MFEM_ABORT("unknown FiniteElementCollection: " << name);
@@ -2544,7 +2541,7 @@ FiniteElementCollection *NURBSFECollection::GetTraceCollection() const
 
 KernelFECollection::KernelFECollection(int D,
                                        int numPointsD,
-                                       int h,
+                                       double h,
                                        RBFFunction *func,
                                        DistanceMetric *dist,
                                        int order)
